+++ robotcode/language_server/common/text_document.py
@@ -216,13 +216,14 @@ class TextDocument:
 
         reference = self.__get_cache_reference(entry)
 
-        with self._lock:
-            e = self._cache[reference]
+        # with self._lock:
+        e = self._cache[reference]
 
-        async with e.lock:
-            if not e.has_data:
-                e.data = await entry(self, *args, **kwargs)
-                e.has_data = True
+        if not e.has_data:
+            async with e.lock:
+                if not e.has_data:
+                    e.data = await entry(self, *args, **kwargs)
+                    e.has_data = True
 
         return cast(_T, e.data)