+++ robotcode/utils/async_cache.py
@@ -27,7 +27,7 @@ class AsyncSimpleLRUCache:
         self._order: Optional[List[Tuple[Any, ...]]] = None
         if self.max_items:
             self._order = []
-        self.lock: Lock = Lock()
+        self._lock: Lock = Lock()
 
     async def has(self, *args: Any, **kwargs: Any) -> bool:
         key = self._make_key(*args, **kwargs)
@@ -40,29 +40,29 @@ class AsyncSimpleLRUCache:
     async def get(self, func: Callable[..., Awaitable[_T]], *args: Any, **kwargs: Any) -> _T:
         key = self._make_key(*args, **kwargs)
 
-        # async with self.lock:
         entry = self._cache[key]
 
-        async with entry.lock:
-            if not entry.has_data:
-                entry.data = await func(*args, **kwargs)
-                entry.has_data = True
+        if not entry.has_data:
+            async with entry.lock:
+                if not entry.has_data:
+                    entry.data = await func(*args, **kwargs)
+                    entry.has_data = True
 
-                if self._order is not None and self.max_items is not None:
-                    async with self.lock:
-                        self._order.insert(0, key)
+                    if self._order is not None and self.max_items is not None:
+                        async with self._lock:
+                            self._order.insert(0, key)
 
-                        if len(self._order) > self.max_items:
-                            del self._cache[self._order.pop()]
+                            if len(self._order) > self.max_items:
+                                del self._cache[self._order.pop()]
 
-            return cast(_T, entry.data)
+        return cast(_T, entry.data)
 
     @staticmethod
     def _make_key(*args: Any, **kwargs: Any) -> Tuple[Any, ...]:
         return (tuple(_freeze(v) for v in args), hash(frozenset({k: _freeze(v) for k, v in kwargs.items()})))
 
     async def clear(self) -> None:
-        async with self.lock:
+        async with self._lock:
             self._cache.clear()
             if self._order is not None:
                 self._order.clear()